+++ .rubocop.yml
@@ -1,4 +1,5 @@
 AllCops:
+  
   NewCops: enable
   Exclude:
     - "Guardfile"
